Extract lib/zendesk_api/resource_tag.rb
